Implement list and delete storage clusters
