+++ api/healthcheck/health_api_StorageClusters.go
@@ -20,7 +20,7 @@ func (api *HealthCheckApi) ListStorageClustersHealth(w http.ResponseWriter, r *h
 	queryParams := map[string]interface{}{
 		"fields": "label,clusterType",
 	}
-	services, res, err := aysClient.Ays.ListServicesByRole("block_cluster", api.AysRepo, nil, queryParams)
+	services, res, err := aysClient.Ays.ListServicesByRole("storagecluster.block", api.AysRepo, nil, queryParams)
 	if !tools.HandleAYSResponse(err, res, w, "listing storagecluster health checks") {
 		return
 	}
@@ -43,7 +43,7 @@ func (api *HealthCheckApi) ListStorageClustersHealth(w http.ResponseWriter, r *h
 		}
 
 		// Get healthcheck data
-		serviceName := fmt.Sprintf("block_cluster_%s", service.Name)
+		serviceName := fmt.Sprintf("storagecluster.block_%s", service.Name)
 		healthService, res, err := aysClient.Ays.GetServiceByName(serviceName, "healthcheck", api.AysRepo, nil, nil)
 		if !tools.HandleAYSResponse(err, res, w, "listing storage clusters health checks") {
 			return
