+++ api/storagecluster/storageclusters_api_ListAllClusters.go
@@ -17,44 +17,31 @@ func (api *StorageclustersAPI) ListAllClusters(w http.ResponseWriter, r *http.Re
 		return
 	}
 	respBody := []string{}
-
-	for _, role := range []string{"block_cluster", "object_cluster"} {
-		clusters, err := getClusters(aysClient, api.AysRepo, role)
-		if err != nil {
-			tools.WriteError(w, http.StatusInternalServerError, err, "")
-			return
-		}
-
-		respBody = append(respBody, clusters...)
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(&respBody)
-}
-
-func getClusters(aysClient *tools.AYStool, aysRepo string, role string) ([]string, error) {
-	respBody := []string{}
-
-	type Data struct {
+	type data struct {
 		Label string `json:"label"`
 	}
-
 	query := map[string]interface{}{
 		"fields": "label",
 	}
-	services, _, err := aysClient.Ays.ListServicesByRole(role, aysRepo, nil, query)
-
+	services, res, err := aysClient.Ays.ListServicesByRole("storagecluster", api.AysRepo, nil, query)
 	if err != nil {
-		return nil, err
+		tools.WriteError(w, http.StatusInternalServerError, err, "Error calling ays list service")
+		return
+	}
+	if res.StatusCode != http.StatusOK {
+		w.WriteHeader(res.StatusCode)
+		return
 	}
 
 	for _, service := range services {
-		data := Data{}
-		if err := json.Unmarshal(service.Data, &data); err != nil {
-			return nil, err
+		Data := data{}
+		if err := json.Unmarshal(service.Data, &Data); err != nil {
+			tools.WriteError(w, http.StatusInternalServerError, err, "Error unmarshaling ays response")
+			return
 		}
-		respBody = append(respBody, data.Label)
+		respBody = append(respBody, Data.Label)
 	}
-	return respBody, nil
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	json.NewEncoder(w).Encode(&respBody)
 }
