+++ pyclient/zeroos/orchestrator/sal/StoragePool.py
@@ -96,8 +96,7 @@ class StoragePool(Mountable):
         Destroy storage pool
         param zero: write zeros (nulls) to the first 500MB of each disk in this storagepool
         """
-        if self.mountpoint:
-            self.umount()
+        self.umount()
         partitionmap = {}
         for disk in self.node.disks.list():
             for partition in disk.partitions:
