+++ raml/types.raml
@@ -1070,12 +1070,9 @@ Cluster:
     nodes:
       type: string[]
       description: List of node ID on which storage server are deployed for this cluster
-    dataStorage:
+    storageServers:
       type: StorageServer[]
       description: List of data storage servers in the cluster
-    metadataStorage:
-      type: StorageServer[]
-      description: List of metadata storage servers in the cluster
 
 # Vdisks API
 VdiskCreate:
