+++ templates/storagecluster.block/config.yaml
