+++ templates/storagecluster.object/config.yaml
