+++ templates/vdisk/actions.py
@@ -43,7 +43,7 @@ def install(job):
             CMD = '/bin/zeroctl copy vdisk --config {etcd} {src_name} {dst_name} {tgtcluster}'
 
             if objectStoragecluster:
-                object_st = service.aysrepo.serviceGet(role='object_cluster', instance=objectStoragecluster)
+                object_st = service.aysrepo.serviceGet(role='storagecluster.object', instance=objectStoragecluster)
                 dataShards = object_st.model.data.dataShards
                 parityShards = object_st.model.data.parityShards
                 CMD += ' --data-shards %s --parity-shards %s' % (dataShards, parityShards)
@@ -148,7 +148,7 @@ def get_cluster_config(job, type="block"):
     vdiskstore = service.parent
 
     cluster = vdiskstore.model.data.blockCluster if type == "block" else vdiskstore.model.data.objectCluster
-    role = "block_cluster" if type == "block" else "object_cluster"
+    role = "storagecluster.block" if type == "block" else "storagecluster.object"
 
     storageclusterservice = service.aysrepo.serviceGet(role=role,
                                                        instance=cluster)
